The summer evening had begun to fold the world in its mysterious embrace.
Far away in the west the sun was setting and the last glow of all too fleeting day lingered lovingly on sea and strand, on the proud promontory of dear old Howth guarding as ever the waters of the bay, on the weedgrown rocks along Sandymount shore and, last but not least, on the quiet church whence there streamed forth at times upon the stillness the voice of prayer to her who is in her pure radiance a beacon ever to the stormtossed heart of man, Mary, star of the sea.

The three girl friends were seated on the rocks, enjoying the evening scene and the air which was fresh but not too chilly. Many a time and oft were they wont to come there to that favourite nook to have a cosy chat beside the sparkling waves and discuss matters feminine, Cissy Caffrey and Edy Boardman with the baby in the pushcar and Tommy and Jacky Caffrey, two little curlyheaded boys, dressed in sailor suits with caps to match and the name H. M. S. Belleisle printed on both. For Tommy and Jacky Caffrey were twins, scarce four years old and very noisy and spoiled twins sometimes but for all that darling little fellows with bright merry faces and endearing ways about them. They were dabbling in the sand with their spades and buckets, building castles as children do, or playing with their big coloured ball, happy as the day was long. And Edy Boardman was rocking the chubby baby to and fro in the pushcar while that young gentleman fairly chuckled with delight. He was but eleven months and nine days old and, though still a tiny toddler, was just beginning to lisp his first babyish words. Cissy Caffrey bent over him to tease his fat little plucks and the dainty dimple in his chin.

— Now, baby, Cissy Caffrey said. Say out big, big. I want a drink of water.

And baby prattled after her :

— A jink a jink a jawbo.

Cissy Caffrey cuddled the wee chap for she was awfully fond of children, so patient with little sufferers and Tommy Caffrey could never be got to take his castor oil unless it was Cissy Caffrey that held his nose and promised him the scatty heel of the loaf of brown bread with golden syrup on. What a persuasive power that girl had! But to be sure baby was as good as gold, a perfect little dote in his new fancy bib. None of your spoilt beauties, Flora Mac Flimsy sort, was Cissy Caffrey. A truerhearted lass never drew the breath of life, always with a laugh in her gipsylike eyes and a frolicsome word on her cherryripe red lips, a girl lovable in the extreme. And Edy Boardman laughed too at the quaint language of little brother.

But just then there was a slight altercation between Master Tommy and Master Jacky. Boys will be boys and our two twins were no exception to this golden rule. The apple of discord was a certain castle of sand which Master Jacky had built and Master Tommy would have it right go wrong that it was to be architecturally improved by a frontdoor like the Martello tower had. But if Master Tommy was headstrong Master Jacky was selfwilled too and, true to the maxim that every little Irishman’s house is his castle, he fell upon his hated rival and to such purpose that the wouldbe assailant came to grief and (alas to relate!) the coveted castle too. Needless to say the cries of discomfited Master Tommy drew the attention of the girl friends.

— Come here, Tommy, his sister called imperatively, at once! And you, Jacky, for shame to throw poor Tommy in the dirty sand. Wait till I catch you for that.

His eyes misty with unshed tears Master Tommy came at her call for their big sister’s word was law with the twins. And in a sad plight he was after his misadventure. His little man-o’-war top and unmentionables were full of sand but Cissy was a past mistress in the art of smoothing over life’s tiny troubles and and very quickly not one speck of sand was to be seen on his smart little suit. Still the blue eyes were glistening with hot tears that would well up so she kissed away the hurtness and shook her hand at Master Jacky the culprit and said if she was near him she wouldn’t be far from him, her eyes dancing in admonition.

— Nasty bold Jacky! she cried.

She put an arm round the little mariner and coaxed winningly :

— What’s your name? Butter and cream?

— Tell us who is your sweetheart, spoke Edy Boardman. Is Cissy your sweetheart?

— Nao, tearful Tommy said.

— Is Edy Boardman your sweetheart? Cissy queried.

— Nao, Tommy said.

— I know, Edy Boardman said none too amiably with an arch glance from her shortsighted eyes. I know who is Tommy’s sweetheart, Gerty is Tommy’s sweetheart.

— Nao, Tommy said on the verge of tears.

Cissy’s quick motherwit guessed what was amiss and she whispered to Edy Boardman to take him there behind the pushcar where the gentlemen couldn’t see and to mind he didn’t wet his new tan shoes.

But who was Gerty?