Deshil Holles Eamus.
Deshil Holles Eamus. Deshil Holles Eamus.

Send us, bright one, light one, Horhorn, quickening and wombfruit. Send us, bright one, light one, Horhorn, quickening and wombfruit. Send us bright one, light one, Horhorn, quickening and wombfruit.

Hoopsa, boyaboy, hoopsa! Hoopsa, boyaboy, hoopsa! Hoopsa, boyaboy, hoopsa.

Universally that person’s acumen is esteemed very little perceptive concerning whatsoever matters are being held as most profitable by mortals with sapience endowed to be studied who is ignorant of that which the most in doctrine erudite and certainly by reason of that in them high mind’s ornament deserving of veneration constantly maintain when by general consent they affirm that other circumstances being equal by no exterior splendour is the prosperity of a nation more efficaciously asserted than by the measure of how far forward may have progressed the tribute of its solicitude for that proliferent continuance which of evils the original if it be absent when fortunately present constitutes the certain sign of omnipollent nature’s incorrupted benefaction. For who is there who anything of some significance has apprehended but is conscious that that exterior splendour may be the surface of a downwardtending lutulent reality or on the contrary anyone so is there inilluminated as not to perceive that as no nature’s boon can contend against the bounty of increase so it behoves every most just citizen to become the exhortator and admonisher of his semblables and to tremble lest what had in the past been by the nation excellently commenced might be in the future not with similar excellence accomplished if an inverecund habit shall have gradually traduced the honourable by ancestors transmitted customs to that thither of profundity that that one was audacious excessively who would have the hardihood to rise affirming that no more odious offence can for anyone be than to oblivious neglect to consign that evangel simultaneously command and promise which on all mortals with prophecy of abundance or with diminution’s menace that exalted of reiteratedly procreating function ever irrevocably enjoined?

It is not why therefore we shall wonder if, as the best historians relate, among the Celts, who nothing that was not in its nature admirable admired, the art of medicine shall have been highly honoured. Not to speak of hostels, leperyards, sweating chambers, plaguegraves, their greatest doctors, the O’Shiels, the O’Hickeys, the O’Lees, have sedulously set down the divers methods by which the sick and the relapsed found again health whether the malady had been the trembling withering or loose boyconnell flux. Certainly in every public work which in it anything of gravity contains preparation should be with importance commensurate and therefore a plan was by them adopted (whether by having preconsidered or as the maturation of experience it is difficult in being said which the discrepant opinions of subsequent inquirers are not up to the present congrued to render manifest) whereby maternity was so far from all accident possibility removed that whatever care the patient in that allhardest of woman hour chiefly required and not solely for the copiously opulent but also for her who not being sufficiently moneyed scarcely and often not even scarcely could subsist valiantly and for an inconsiderable emolument was provided.

To her nothing already then and thenceforward was anyway able to be molestful for this chiefly felt all citizens except with proliferent mothers prosperity at all not to can be and as they had received eternity gods mortals generation to befit them her beholding, when the case was so having itself, parturient in vehicle thereward carrying desire immense among all one another was impelling on of her to be received into that domicile. O thing of prudent nation not merely in being seen but also even in being related worthy of being praised that they her by anticipation went seeing mother, that she by them suddenly to be about to be cherished had been begun she felt!