Preparatory to anything else Mr Bloom brushed off the greater bulk of the shavings
And handed Stephen the hat and ashplant and bucked him up generally in orthodox Samaritan fashion, which he very badly needed. His (Stephen’s) mind was not exactly what you would call wandering but a bit unsteady and on his expressed desire for some beverage to drink Mr Bloom, in view of the hour it was and there being no pumps of Vartry water available for their ablutions, let alone drinking purposes, hit upon an expedient by suggesting, off the reel, the propriety of the cabman’s shelter, as it was called, hardly a stonesthrow away near Butt Bridge where they might hit upon some drinkables in the shape of a milk and soda or a mineral. But how to get there was the rub. For the nonce he was rather nonplussed but inasmuch as the duty plainly devolved upon him to take some measures on the subject he pondered suitable ways and means during which Stephen repeatedly yawned. So far as he could see he was rather pale in the face so that it occurred to him as highly advisable to get a conveyance of some description which would answer in their then condition, both of them being e. d. ed, particularly Stephen, always assuming that there was such a thing to be found. Accordingly, after a few such preliminaries, as, in spite of his having forgotten to take up his rather soapsuddy handkerchief after it had done yeoman service in the shaving line, brushing they both walked together along Beaver street, or, more properly, lane, as far as the farrier’s and the distinctly fetid atmosphere of the livery stables at the corner of Montgomery street where they made tracks to the left from thence debouching into Amiens Street round by the corner of Dan Bergin’s. But, as he confidently anticipated, there was not a sign of a Jehu plying for hire anywhere to be seen except a fourwheeler, probably engaged by some fellows inside on the spree, outside the North Star Hotel and there was no symptom of its budging a quarter of an inch when Mr Bloom, who was anything but a professional whistler, endeavoured to hail it by emitting a kind of a whistle, holding his arms arched over his head, twice. This was a quandary but, bringing commonsense to bear on it, evidently there was nothing for it but put a good face on the matter and foot it which they accordingly did. So, bevelling around by Mullet's and the Signal House, which they shortly reached, they proceeded perforce in the direction of Amiens Street railway terminus, Mr Bloom being handicapped by the circumstance that one of the back buttons of his trousers had, to vary the timehonoured adage, gone the way of all buttons, though, entering thoroughly into the spirit of the thing, he heroically made light of the mischance. So as neither of them were particularly pressed for time, as it happened, and the temperature refreshing since it cleared up after the recent visitation of Jupiter Pluvius, they dandered along past by where the empty vehicle was waiting without a fare or a jarvey. As it so happened a Dublin United Tramways Company's sandstrewer happening to be returning and the elder man recounted to his companion à propos of the incident his own truly miraculous escape of some little while back. They passed the main entrance of the Great Northern railway station, the starting point for Belfast, where of course all traffic was suspended at that late hour, and passing the back door of the morgue (a not very enticing locality, not to say gruesome to a degree, more especially at night), ultimately gained the Dock Tavern and in due course turned into Store street, famous for its C division police station. Between this point and the high, at present unlit, warehouses of Beresford Place Stephen thought to think of Ibsen, associated with Baird's, the stonecutter's in his mind somehow in Talbot Place, first turning on the right, while the other, who was acting as his fidus Achates inhaled with internal satisfaction the smell of James Rourke's city bakery, situated quite close to where they were, the very palatable odour indeed of our daily bread, of all commodities of the public the primary and most indispensable. Bread, the staff of life, earn your bread, O tell me where is fancy bread? At Rourke's the baker's, it is said.